+++ packages/tracing/src/browser/web-vitals/getFID.ts
@@ -15,75 +15,32 @@
  */
 
 import { bindReporter } from './lib/bindReporter';
-import { getFirstHidden } from './lib/getFirstHidden';
+import { getVisibilityWatcher } from './lib/getVisibilityWatcher';
 import { initMetric } from './lib/initMetric';
 import { observe, PerformanceEntryHandler } from './lib/observe';
 import { onHidden } from './lib/onHidden';
-import { ReportHandler } from './types';
+import { PerformanceEventTiming, ReportHandler } from './types';
 
-interface FIDPolyfillCallback {
-  (value: number, event: Event): void;
-}
-
-interface FIDPolyfill {
-  onFirstInputDelay: (onReport: FIDPolyfillCallback) => void;
-}
-
-declare global {
-  interface Window {
-    perfMetrics: FIDPolyfill;
-  }
-}
-
-// https://wicg.github.io/event-timing/#sec-performance-event-timing
-interface PerformanceEventTiming extends PerformanceEntry {
-  processingStart: DOMHighResTimeStamp;
-  cancelable?: boolean;
-  target?: Element;
-}
-
-export const getFID = (onReport: ReportHandler): void => {
+export const getFID = (onReport: ReportHandler, reportAllChanges?: boolean): void => {
+  const visibilityWatcher = getVisibilityWatcher();
   const metric = initMetric('FID');
-  const firstHidden = getFirstHidden();
+  let report: ReturnType<typeof bindReporter>;
 
   const entryHandler = (entry: PerformanceEventTiming): void => {
     // Only report if the page wasn't hidden prior to the first input.
-    if (entry.startTime < firstHidden.timeStamp) {
+    if (report && entry.startTime < visibilityWatcher.firstHiddenTime) {
       metric.value = entry.processingStart - entry.startTime;
       metric.entries.push(entry);
-      metric.isFinal = true;
-      report();
+      report(true);
     }
   };
 
   const po = observe('first-input', entryHandler as PerformanceEntryHandler);
-  const report = bindReporter(onReport, metric, po);
-
   if (po) {
+    report = bindReporter(onReport, metric, reportAllChanges);
     onHidden(() => {
       po.takeRecords().map(entryHandler as PerformanceEntryHandler);
       po.disconnect();
     }, true);
-  } else {
-    if (window.perfMetrics && window.perfMetrics.onFirstInputDelay) {
-      window.perfMetrics.onFirstInputDelay((value: number, event: Event) => {
-        // Only report if the page wasn't hidden prior to the first input.
-        if (event.timeStamp < firstHidden.timeStamp) {
-          metric.value = value;
-          metric.isFinal = true;
-          metric.entries = [
-            {
-              entryType: 'first-input',
-              name: event.type,
-              target: event.target,
-              cancelable: event.cancelable,
-              startTime: event.timeStamp,
-              processingStart: event.timeStamp + value,
-            } as PerformanceEventTiming,
-          ];
-          report();
-        }
-      });
-    }
   }
 };