+++ packages/vue/src/index.ts
@@ -1,4 +1,6 @@
 export * from '@sentry/browser';
 
 export { init } from './sdk';
-export { vueRouterInstrumentation } from './vuerouter';
+export { vueRouterInstrumentation } from './router';
+export { attachErrorHandler } from './errorhandler';
+export { createTracingMixins } from './tracing';